Mi proyecto/web/requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
djangorestframework==3.14.0